+++ Changes.md
@@ -1,3 +1,8 @@
+#### Release 0.0.15
+- add a param converters in as_dict function, which means u can do the convertion according to your requirements. the format for this param is {'key': func}
+- add bulk_query decorator, so you can insert multiple rows to the db
+- add as_df function to the RecordCollection
+
 #### Release 0.0.14
 - Fix the initialize issue in the way like FlaskRecords()
 - Modified the error in the Doc